refactor: split Downloader implementation
