+++ crates/serverless/src/deployments/downloader/s3_bucket.rs
@@ -2,10 +2,7 @@ use anyhow::Result;
 use async_trait::async_trait;
 use s3::Bucket;
 
-#[async_trait]
-pub trait Downloader: Sync + Clone {
-    async fn download(&self, path: String) -> Result<Vec<u8>>;
-}
+use super::Downloader;
 
 pub struct S3BucketDownloader {
     bucket: Bucket,
@@ -30,20 +27,5 @@ impl Downloader for S3BucketDownloader {
     async fn download(&self, path: String) -> Result<Vec<u8>> {
         let object = self.bucket.get_object(path).await?;
         Ok(object.bytes().to_vec())
-    }
-}
-
-pub struct FakeDownloader;
-
-#[async_trait]
-impl Downloader for FakeDownloader {
-    async fn download(&self, _path: String) -> Result<Vec<u8>> {
-        Ok(vec![])
-    }
-}
-
-impl Clone for FakeDownloader {
-    fn clone(&self) -> Self {
-        Self {}
     }
 }
